requirements.txt
python-telegram-bot>=21.0
yt-dlp>=2024.01.01
ffmpeg-python>=0.2.0
httpx>=0.27.0
pydantic-settings>=2.0
loguru>=0.7.0
python-dotenv>=1.0.0
